Windowing stable. Updated to work for 32-bit machines with long long

+++ cutfunction.hpp
@@ -37,37 +37,37 @@
 class CutFunction{
 	private:
 		//Output Truth Table, Circuit Function
-		std::map<unsigned long, std::string> m_HashTable;
+		std::map<unsigned long long, std::string> m_HashTable;
 
 		//Output Truth Table, count 
-		std::map<unsigned long, int> m_FunctionCount;
+		std::map<unsigned long long, int> m_FunctionCount;
 
 
 		//Function, Nodes with the function output
-		std::map<unsigned long, std::set<unsigned> > m_NodeFunction;
-		//std::map<unsigned long, std::map<unsigned,unsigned> >m_CutInputIndex;
+		std::map<unsigned long long, std::set<unsigned> > m_NodeFunction;
+		//std::map<unsigned long long, std::map<unsigned,unsigned> >m_CutInputIndex;
 
 
 		//Function, List of inputs that have that specific output
 		//Output node is stored as the last node in inner vector
-		std::map<unsigned long, std::vector<std::vector<unsigned>*> >m_PortMap; 
+		std::map<unsigned long long, std::vector<std::vector<unsigned>*> >m_PortMap; 
 		std::map<std::string, unsigned int> m_PrimInputSize;
 
 
 		//Set of functions that define the output, list of input and output of the cut
-		std::map<std::set<unsigned long>, std::vector<std::vector<unsigned> > > m_UniqueFunction;
+		std::map<std::set<unsigned long long>, std::vector<std::vector<unsigned> > > m_UniqueFunction;
 
 
 		//AIG Node, Output Tru
-		unsigned long* m_Xval;
+		unsigned long long* m_Xval;
 
 		CutEnumeration* m_CutEnumeration;
 		AIG* m_AIG;
 
 	public:
 		//AIG Node, Truth Table Result
-		std::map<unsigned int, unsigned long> m_NodeValue;
-		std::map<unsigned int, unsigned long> m_NodeValueIn;
+		std::map<unsigned int, unsigned long long> m_NodeValue;
+		std::map<unsigned int, unsigned long long> m_NodeValueIn;
 
 
 		//Constructors
@@ -80,9 +80,9 @@ class CutFunction{
 		void preProcessLibrary(std::string);
 		void processAIGCuts(bool);
 		void processAIGCuts_Perm(bool);
-		//void processLUTs(Graph*, std::map<std::string, std::set<unsigned long> >&);
-		unsigned long calculate(unsigned);
-		unsigned long calculate2(unsigned);
+		//void processLUTs(Graph*, std::map<std::string, std::set<unsigned long long> >&);
+		unsigned long long calculate(unsigned);
+		unsigned long long calculate2(unsigned);
 		unsigned int* setPermutation(int);
 
 
@@ -92,10 +92,10 @@ class CutFunction{
 
 
 		//Getters
-		void getPortMap(std::map<unsigned long, std::vector<std::vector<unsigned>*> >& );
-		void getHashMap(std::map<unsigned long, std::string>&);
+		void getPortMap(std::map<unsigned long long, std::vector<std::vector<unsigned>*> >& );
+		void getHashMap(std::map<unsigned long long, std::string>&);
 		void setParams(CutEnumeration*, AIG*);
-		void getFunctionCount(std::map<unsigned long, int>& );
+		void getFunctionCount(std::map<unsigned long long, int>& );
 
 
 		//Print Functions
