+++ vertex.cpp
@@ -71,7 +71,7 @@ void Vertex::setLevel(int level){
 
 void Vertex::setLUT(std::string func){
 	m_LUT = 0;
-	unsigned long base = 1;
+	unsigned long long base = 1;
 	//Convert function string into an integer
 	for(int i = func.size()-1; i >= 0; i--){
 		if( func[i] <= '9'){
@@ -88,11 +88,11 @@ void Vertex::setLUT(std::string func){
 }
 
 
-void Vertex::setLUT(unsigned long lut){
+void Vertex::setLUT(unsigned long long lut){
 	m_LUT = lut;
 }
 
-unsigned long Vertex::getLUT(){
+unsigned long long Vertex::getLUT(){
 	return m_LUT;
 }
