+++ vertex.hpp
@@ -25,7 +25,7 @@ class Vertex
 		std::string m_Name;
 		int m_ID;    //Unique ID assigned
 		int m_Level;
-		unsigned long m_LUT;
+		unsigned long long m_LUT;
 
 		//Set of vertices connected into the current vertex
 		std::vector<Vertex*> m_Input;
@@ -42,12 +42,12 @@ class Vertex
 
 		void setLevel(int);
 		void setLUT(std::string);
-		void setLUT(unsigned long);
+		void setLUT(unsigned long long);
 		void setType(std::string);
 		void setName(std::string);
 		void setVertexID(int);
 
-		unsigned long getLUT();
+		unsigned long long getLUT();
 		unsigned int getOVSize();
 		unsigned int getIVSize();
 		int getID();
